+++ docs/versionhistory.rst
@@ -3,6 +3,15 @@ Version history
 
 This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
+**UNRELEASED**
+
+- The following functions and methods have been moved:
+
+  * ``current_default_worker_thread_limiter`` → ``anyio.to_thread.current_default_thread_limiter``
+  * ``run_sync_in_worker_thread()`` → ``anyio.to_thread.run_sync()``
+  * ``run_async_from_thread()`` → ``anyio.from_thread.run()``
+  * ``run_sync_from_thread()`` → ``anyio.from_thread.run_sync()``
+
 **3.0.0rc2**
 
 - Changed the asyncio ``CancelScope`` to raise a ``RuntimeError`` if cancel scopes are being exited